Nurse can view patients and record vital signs

+++ main/app/Modules/Appointment/Routes/web.php
@@ -1,7 +1,8 @@
 <?php
 
-use App\Modules\Appointment\Http\Controllers\AppointmentController;
 use Illuminate\Support\Facades\Route;
+use App\Modules\Nurse\Http\Controllers\VitalsController;
+use App\Modules\Appointment\Http\Controllers\AppointmentController;
 
 /*
 |--------------------------------------------------------------------------
@@ -15,8 +16,11 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::prefix('appointments')->name('appointments.')->group(function() {
-    Route::get('/{date?}', [AppointmentController::class, 'index'])->name('index');
+    Route::get('/{date?}', [AppointmentController::class, 'index'])->name('index')->middleware('auth:front_desk_user');
     Route::post('{patient}/create', [AppointmentController::class, 'store'])->name('create')->middleware('auth:front_desk_user');
-    Route::put('{appointment}/post-for-vitals', [AppointmentController::class, 'postForVitals'])->name('post_for_vitals');
-    Route::delete('{appointment}/cancel', [AppointmentController::class, 'destroy'])->name('delete');
-});
+    Route::put('{appointment}/post-for-vitals', [AppointmentController::class, 'postForVitals'])->name('post_for_vitals')->middleware('auth:front_desk_user');
+    Route::delete('{appointment}/cancel', [AppointmentController::class, 'destroy'])->name('delete')->middleware('auth:front_desk_user');
+
+
+    Route::post('{appointment}/vitals/create', [VitalsController::class, 'store'])->name('vitals.create')->middleware('auth:nurse');
+  });
